requirements.txt
Django==4.2.*
django-auth-ldap
django-cors-headers
django-datatable-view
django-extensions
django-filter
django-modeladmin-reorder
django-loginas
django-simple-history
django-ckeditor
djangorestframework
#djangorestframework-datatables
#djangorestframework-filters==1.0.0.dev0
djangorestframework-queryfields
whitenoise
gevent
gunicorn
jira-python
psycopg2-binary==2.9.5
pillow
scalpl
iso3166
docker==6.1.3

git+https://github.com/den-gts/django-database-files-3000.git
git+https://github.com/perfguru87/django-atlassian.git

